janv 27

+++ include/core/glwe/glwe_ct_params.h
@@ -11,7 +11,7 @@ typedef struct glwe_ct_params {
   // ct(a,b)
   uint64_t n_limbs;   // (k+1)l or (k+1)l - 1 if l_a and l_b are different l, 
                       // at each limb we have a polynomial of degree N (Zn[X])
-  uint64_t sigma;     // The standard deviation of the error distribution.                
+  double sigma;     // The standard deviation of the error distribution.                
 } GLWECtParams;
 
 /**
@@ -24,7 +24,7 @@ typedef struct glwe_ct_params {
  * @param sigma The standard deviation of the error distribution.
  * @return GLWECtParams* 
  */
-GLWECtParams* new_glwe_ct_params(uint64_t N, uint64_t k, uint64_t kappa, uint64_t n_limbs, uint64_t sigma);
+GLWECtParams* new_glwe_ct_params(uint64_t N, uint64_t k, uint64_t kappa, uint64_t n_limbs, double sigma);
 
 /**
  * @brief Deletes the set of GLWE parameters.
